Make Slf4jLogger no longer crash when its used with the Logger[F] type
This solution fixes a problem in the previous commit where the program would
crash with ??? when the logger was being referenced from the `Logger[F]` interface.

What this implementation effectively achives is that when the user uses the the the
slf4j backend wrapper via the subtype io.chrisdavenport.log4cats.slf4j.Slf4jLogger
then they get macro-based optimizations like you see in iolog4s. But when they switch
to the abstract algebra of `Logger[F]` then they lose this macro-based optimization
and they get something that is equivalent to the old log4cats wrapper.