Add LoggerFactory typeclass
The reasoning for this commit is that I want to avoid having a Sync instance just for being able to log in a http4s api.

Extract LoggerName from macro and make it an implicit.
This makes it easier to have forwarders to Slf4jLogger.

This should be binary compatible, I had to add some exclusions to mima because of the missing macro methods.